properly handle partial buffers in the (sub)partition queues
Make use of DuplicatedNetworkBuffer instances to have shared-content but
indices-independent duplicates so that multiple readers are handled correctly,
e.g. for events. Also, now, we define the transferred content once the buffer
is polled from the queue so that any statistics are correct.

TODO: Aside from some optimisations which are left to implement, spilled buffers
currently do not store the reader and writer indexes and their contents are thus
transmitted multiple times!